+++ src/KappaDuck.Quack/Video/Pixels/ColorF.cs
@@ -5,7 +5,7 @@ using KappaDuck.Quack.Geometry;
 using System.Numerics;
 using System.Text.Unicode;
 
-namespace KappaDuck.Quack.Video;
+namespace KappaDuck.Quack.Video.Pixels;
 
 /// <summary>
 /// Represents an RGBA color with floating-point channels, normally in the range 0-1.
