+++ src/KappaDuck.Quack/Video/Pixels/Colors.cs
@@ -1,7 +1,7 @@
 // Copyright (c) KappaDuck.
 // Licensed under the MIT license.
 
-namespace KappaDuck.Quack.Video;
+namespace KappaDuck.Quack.Video.Pixels;
 
 /// <summary>
 /// Provides the standard set of named <see cref="Color"/> values.
