+++ src/KappaDuck.Quack/Video/Pixels/PixelFormat.cs
@@ -1,7 +1,7 @@
 // Copyright (c) KappaDuck.
 // Licensed under the MIT license.
 
-namespace KappaDuck.Quack.Video;
+namespace KappaDuck.Quack.Video.Pixels;
 
 /// <summary>
 /// Identifies the in-memory layout of a pixel.
